SuSEFirewall: Move SuSEFirewall{,2,d} classes to new files
Yast2 coding standard requires each class to be on a separate file.
As such move SuSEFirewall2, SuSEFirewalld and SuSEFirewall classes
to separate files. These files are being moved to the 'lib/network'
subdirectory as well. Moreover, add a new rubocop configuration file
to silence warnings about variables and methods since these are still
part of the public API.
