+++ library/network/src/lib/network/.rubocop.yml
@@ -1,0 +1,21 @@
+# use the shared Yast defaults
+inherit_from:
+  - ../../../../../.rubocop.yml
+
+Style/MethodName:
+  Exclude:
+    - "susefirewall.rb"
+    - "susefirewalld.rb"
+    - "susefirewall2.rb"
+
+Style/Next:
+  # skip this check for SuSEFirewall modules since it forces us to
+  # refactor the code in a sub-optimal way.
+  Exclude:
+    - "susefirewalld.rb"
+
+Style/VariableName:
+  Exclude:
+    - "susefirewall.rb"
+    - "susefirewalld.rb"
+    - "susefirewall2.rb"
